Did something
Split testing and real placement and made place object universal
