+++ Assets/Scripts/AssetPlacementTesting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 60fe18513469d9943ac30d8d776b2423
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
